Move _generate_coordinate_range to different utils module
